The TowerTAPE filing system now checks to see what file type it is loading.
This means that it now displays what type it is and doesn't load the wrong type.

Also, there is a new field in the header at the start which must contain 'HEAD'.

This is to prevent misidentification of other byte streams as headers.

Other minor tidy-up operations have also occurred.